refactor: extract tokenizer
