+++ src/ByteSync.Client/Interfaces/Services/Filtering/IFilterTokenizer.cs
@@ -1,0 +1,10 @@
+﻿using ByteSync.Business.Filtering.Parsing;
+
+namespace ByteSync.Interfaces.Services.Filtering;
+
+public interface IFilterTokenizer
+{
+    void Initialize(string filterText);
+    
+    FilterToken GetNextToken();
+}
